Avoid double parse in literal detection
The last commit introduces literal detection that used a int
literal parser to check whether the given value was an int
literal and fell back to a double parser when no int was
found.

This can lead to double parsing, for instance when a string
starts with digits, or for all doubles. (Which would be
parsed as integers until the dot, then we would do the same
work again in the double parser).

This adds a function, that is guaranteed to only do one pass
over the data.