Fix emitter::emit(event::event&&)
The constructor is meant to be taking temporaries. Before
this change, it would try to move the rvalue reference into
a local event. This caused problems on some platforms.

Now we just call libyaml's emit with the rvalue ref, since
it is totally OK in this case if the temporary gets
destroyed after the call.