Bugfix where some unqoted literals would be quoted
This could happen when such a literal (int/float/bool or null)
occurred after an odd number of arrays or objects within an objects.

The reason for this was that we keep track of keys/values in objects
so we can make sure that the keys are always quoted.
We forgot to count arrays and objects as keys, so this would mess up
our counting.